+++ src/App.tsx
@@ -12,6 +12,7 @@ import {
   PatientConditions,
   PatientConditionsOutside,
   PatientDocuments,
+  PatientImmunizations,
   PatientMedications,
   PatientMedicationsOutside,
   PatientProvider,
@@ -82,6 +83,11 @@ const components: DemoComponent[] = [
     title: "Patient Documents",
   },
   {
+    name: "immunizations",
+    render: () => <PatientImmunizations />,
+    title: "Patient Immunizations",
+  },
+  {
     name: "medications",
     render: () => <PatientMedications />,
     title: "Patient Medications",
@@ -110,14 +116,11 @@ const components: DemoComponent[] = [
           "allergies",
           "conditions",
           "conditions-outside",
-          "medications",
-          "medications-outside",
-          "observations",
-          "timeline",
-          "observations-outside",
-          "care-team",
           "documents",
           "immunizations",
+          "care-team",
+          "medications",
+          "medications-outside",
         ]}
         title="ZAP"
       />