+++ src/components/content/document/patient-documents.tsx
@@ -1,5 +1,5 @@
 import cx from "classnames";
-import { useRef } from "react";
+import { useMemo, useRef } from "react";
 import { patientDocumentColumns } from "./helpers/columns";
 import { getDateRangeView } from "../resource/helpers/view-date-range";
 import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
@@ -8,21 +8,26 @@ import { ResourceTableActions } from "../resource/resource-table-actions";
 import { EmptyTable } from "@/components/core/empty-table";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { useCTW } from "@/components/core/providers/use-ctw";
-import { usePatientDocument } from "@/fhir/document";
+import { RowActionsProps } from "@/components/core/table/table";
+import { getBinaryDocument } from "@/fhir/binaries";
+import { usePatientDocuments } from "@/fhir/document";
 import { DocumentModel } from "@/fhir/models/document";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
 import { useFQSFeatureToggle } from "@/hooks/use-fqs-feature-toggle";
+import { useBaseTranslations } from "@/i18n";
 
-export type PatientDocumentProps = {
+export type PatientDocumentsProps = {
   className?: string;
+  onAddToRecord?: (document: DocumentModel, binary: fhir4.Binary) => void;
 };
 
-function PatientDocumentsComponent({ className }: PatientDocumentProps) {
+function PatientDocumentsComponent({ className, onAddToRecord }: PatientDocumentsProps) {
   const containerRef = useRef<HTMLDivElement>(null);
   const { featureFlags } = useCTW();
   const { enabled } = useFQSFeatureToggle("documents");
 
-  const patientDocumentQuery = usePatientDocument();
+  const patientDocumentQuery = usePatientDocuments();
+  const rowActions = useMemo(() => getRowActions({ onAddToRecord }), [onAddToRecord]);
   const { viewOptions, defaultView } = getDateRangeView<DocumentModel>("dateCreated");
   const { data, setViewOption } = useFilteredSortedData({
     defaultView,
@@ -59,10 +64,45 @@ function PatientDocumentsComponent({ className }: PatientDocumentProps) {
         }
         columns={patientDocumentColumns(featureFlags?.enableViewFhirButton)}
         onRowClick={openDetails}
+        rowActions={rowActions}
       />
     </div>
   );
 }
+
+type ExtraRowActionProps = {
+  onAddToRecord?: (document: DocumentModel, binary: fhir4.Binary) => void;
+};
+
+const getRowActions =
+  ({ onAddToRecord }: ExtraRowActionProps) =>
+  (props: RowActionsProps<DocumentModel>) =>
+    <RowActions {...props} onAddToRecord={onAddToRecord} />;
+
+type RowActionsProps2 = RowActionsProps<DocumentModel> & ExtraRowActionProps;
+
+const RowActions = ({ record, onAddToRecord }: RowActionsProps2) => {
+  const { t } = useBaseTranslations();
+  const { getRequestContext } = useCTW();
+  const { binaryId } = record;
+
+  return onAddToRecord && binaryId ? (
+    <div className="ctw-flex ctw-space-x-2">
+      <button
+        type="button"
+        className="ctw-btn-primary ctw-ml-1 ctw-capitalize"
+        data-zus-telemetry-click="Add to record"
+        data-testid="add-to-record"
+        onClick={async () => {
+          const binary = await getBinaryDocument(await getRequestContext(), binaryId);
+          onAddToRecord(record, binary);
+        }}
+      >
+        {t("resourceTable.add")}
+      </button>
+    </div>
+  ) : null;
+};
 
 export const PatientDocuments = withErrorBoundary(PatientDocumentsComponent, "PatientDocuments");