+++ src/fhir/medications.ts
@@ -42,8 +42,7 @@ import { MedicationModel } from "@/fhir/models/medication";
 import { MedicationStatementModel } from "@/fhir/models/medication-statement";
 import { PatientModel } from "@/fhir/models/patient";
 import { filterResourcesByBuilderId } from "@/services/common";
-import { createGraphqlClient } from "@/services/fqs/client";
-import { graphQLToFHIR } from "@/services/fqs/graphql-to-fhir";
+import { createGraphqlClient, fqsRequest } from "@/services/fqs/client";
 import {
   MedicationAdministrationGraphqlResponse,
   medicationAdministrationQuery,
@@ -61,11 +60,12 @@ import {
   medicationStatementQuery,
 } from "@/services/fqs/queries/medication-statements";
 import { errorResponse } from "@/utils/errors";
-import { cloneDeep, uniqWith } from "@/utils/nodash";
+import { cloneDeep, sortBy, uniqWith } from "@/utils/nodash";
 import {
   compact,
   filter,
   find,
+  sortBy as fpSortBy,
   get,
   groupBy,
   last,
@@ -74,7 +74,6 @@ import {
   omit,
   pipe,
   propEq,
-  sortBy,
   split,
 } from "@/utils/nodash/fp";
 import { QUERY_KEY_MEDICATION_HISTORY } from "@/utils/query-keys";
@@ -161,24 +160,27 @@ export async function getBuilderMedicationStatementsFQS(
       informationSourceNot: "Patient", // exclude medication statements where the patient is the information source
     } as MedicationFilter;
     const graphClient = createGraphqlClient(requestContext);
-    const data = await graphClient.request(medicationStatementQuery, {
-      upid: patient.UPID,
-      cursor: "",
-      first: 1000,
-      sort: {
-        lastUpdated: "DESC",
-      },
-      filter: {
-        tag: {
-          nonematch: [SYSTEM_SUMMARY, `${SYSTEM_ZUS_THIRD_PARTY}`],
-          // TODO: There's a bug in FQS that doesn't allow filtering with nonematch AND allmatch.
-          // Uncomment the line below once https://zeushealth.atlassian.net/browse/DRT-249 is resolved.
-          // allmatch: [`${SYSTEM_ZUS_OWNER}|builder/${requestContext.builderId}`],
+    const { data } = await fqsRequest<MedicationStatementGraphqlResponse>(
+      graphClient,
+      medicationStatementQuery,
+      {
+        upid: patient.UPID,
+        cursor: "",
+        first: 1000,
+        sort: {
+          lastUpdated: "DESC",
         },
-      },
-    });
-    const cleanData = graphQLToFHIR(data) as MedicationStatementGraphqlResponse;
-    let nodes = cleanData.MedicationStatementConnection.edges.map((x) => x.node);
+        filter: {
+          tag: {
+            nonematch: [SYSTEM_SUMMARY, `${SYSTEM_ZUS_THIRD_PARTY}`],
+            // TODO: There's a bug in FQS that doesn't allow filtering with nonematch AND allmatch.
+            // Uncomment the line below once https://zeushealth.atlassian.net/browse/DRT-249 is resolved.
+            // allmatch: [`${SYSTEM_ZUS_OWNER}|builder/${requestContext.builderId}`],
+          },
+        },
+      }
+    );
+    let nodes = data.MedicationStatementConnection.edges.map((x) => x.node);
     nodes = filterResourcesByBuilderId(
       nodes,
       requestContext.contextBuilderId || requestContext.builderId
@@ -209,21 +211,24 @@ export async function getMedicationStatementsForPatientByIdFQS(
       return { bundle: undefined, medications: [], basic: [] };
     }
     const graphClient = createGraphqlClient(requestContext);
-    const data = await graphClient.request(medicationStatementQuery, {
-      upid: patient.UPID,
-      cursor: "",
-      first: 1000,
-      sort: {
-        lastUpdated: "DESC",
-      },
-      filter: {
-        ids: {
-          anymatch: resourceIds,
+    const { data } = await fqsRequest<MedicationStatementGraphqlResponse>(
+      graphClient,
+      medicationStatementQuery,
+      {
+        upid: patient.UPID,
+        cursor: "",
+        first: 1000,
+        sort: {
+          lastUpdated: "DESC",
         },
-      },
-    });
-    const cleanData = graphQLToFHIR(data) as MedicationStatementGraphqlResponse;
-    const nodes = cleanData.MedicationStatementConnection.edges.map((x) => x.node);
+        filter: {
+          ids: {
+            anymatch: resourceIds,
+          },
+        },
+      }
+    );
+    const nodes = data.MedicationStatementConnection.edges.map((x) => x.node);
     return { bundle: undefined, medications: nodes, basic: [] };
   } catch (e) {
     throw Telemetry.logError(
@@ -243,19 +248,23 @@ export async function getMedicationAdministrationsForPatientByIdFQS(
       return [];
     }
     const graphClient = createGraphqlClient(requestContext);
-    const data = (await graphClient.request(medicationAdministrationQuery, {
-      upid: patient.UPID,
-      cursor: "",
-      first: 1000,
-      sort: {
-        lastUpdated: "DESC",
-      },
-      filter: {
-        ids: {
-          anymatch: resourceIds,
+    const { data } = await fqsRequest<MedicationAdministrationGraphqlResponse>(
+      graphClient,
+      medicationAdministrationQuery,
+      {
+        upid: patient.UPID,
+        cursor: "",
+        first: 1000,
+        sort: {
+          lastUpdated: "DESC",
         },
-      },
-    })) as MedicationAdministrationGraphqlResponse;
+        filter: {
+          ids: {
+            anymatch: resourceIds,
+          },
+        },
+      }
+    );
     const nodes = data.MedicationAdministrationConnection.edges.map((x) => x.node);
     return nodes;
   } catch (e) {
@@ -276,21 +285,25 @@ export async function getMedicationDispensesForPatientByIdFQS(
       return [];
     }
     const graphClient = createGraphqlClient(requestContext);
-    const data = await graphClient.request(medicationDispenseQuery, {
-      upid: patient.UPID,
-      cursor: "",
-      first: 1000,
-      sort: {
-        lastUpdated: "DESC",
-      },
-      filter: {
-        ids: {
-          anymatch: resourceIds,
+    const { data } = await fqsRequest<MedicationDispenseGraphqlResponse>(
+      graphClient,
+      medicationDispenseQuery,
+      {
+        upid: patient.UPID,
+        cursor: "",
+        first: 1000,
+        sort: {
+          lastUpdated: "DESC",
         },
-      },
-    });
-    const cleanData = graphQLToFHIR(data) as MedicationDispenseGraphqlResponse;
-    const nodes = cleanData.MedicationDispenseConnection.edges.map((x) => x.node);
+        filter: {
+          ids: {
+            anymatch: resourceIds,
+          },
+        },
+      }
+    );
+
+    const nodes = data.MedicationDispenseConnection.edges.map((x) => x.node);
     return nodes;
   } catch (e) {
     throw Telemetry.logError(
@@ -310,21 +323,25 @@ export async function getMedicationRequestsForPatientByIdFQS(
       return [];
     }
     const graphClient = createGraphqlClient(requestContext);
-    const data = await graphClient.request(medicationRequestQuery, {
-      upid: patient.UPID,
-      cursor: "",
-      first: 1000,
-      sort: {
-        lastUpdated: "DESC",
-      },
-      filter: {
-        ids: {
-          anymatch: resourceIds,
+
+    const { data } = await fqsRequest<MedicationRequestGraphqlResponse>(
+      graphClient,
+      medicationRequestQuery,
+      {
+        upid: patient.UPID,
+        cursor: "",
+        first: 1000,
+        sort: {
+          lastUpdated: "DESC",
         },
-      },
-    });
-    const cleanData = graphQLToFHIR(data) as MedicationRequestGraphqlResponse;
-    const nodes = cleanData.MedicationRequestConnection.edges.map((x) => x.node);
+        filter: {
+          ids: {
+            anymatch: resourceIds,
+          },
+        },
+      }
+    );
+    const nodes = data.MedicationRequestConnection.edges.map((x) => x.node);
     return nodes;
   } catch (e) {
     throw Telemetry.logError(
@@ -465,24 +482,27 @@ export async function getActiveMedicationsFQS(
 ): Promise<MedicationResults> {
   try {
     const graphClient = createGraphqlClient(requestContext);
-    const data = await graphClient.request(medicationStatementQuery, {
-      upid: patient.UPID,
-      cursor: "",
-      first: 1000,
-      sort: {
-        lastUpdated: "DESC",
-      },
-      filter: {
-        tag: {
-          allmatch: [
-            `${SYSTEM_ZUS_SUMMARY}|Common`,
-            `${SYSTEM_ZUS_OWNER}|builder/${getLensBuilderId(requestContext.env)}`,
-          ],
+    const { data } = await fqsRequest<MedicationStatementGraphqlResponse>(
+      graphClient,
+      medicationStatementQuery,
+      {
+        upid: patient.UPID,
+        cursor: "",
+        first: 1000,
+        sort: {
+          lastUpdated: "DESC",
         },
-      },
-    });
-    const cleanData = graphQLToFHIR(data) as MedicationStatementGraphqlResponse;
-    const nodes = cleanData.MedicationStatementConnection.edges.map((x) => x.node);
+        filter: {
+          tag: {
+            allmatch: [
+              `${SYSTEM_ZUS_SUMMARY}|Common`,
+              `${SYSTEM_ZUS_OWNER}|builder/${getLensBuilderId(requestContext.env)}`,
+            ],
+          },
+        },
+      }
+    );
+    const nodes = data.MedicationStatementConnection.edges.map((x) => x.node);
     const models = setupMedicationStatementModelsWithFQS(nodes);
     if (models.length === 0) {
       Telemetry.countMetric("req.count.active_medications.none", 1, ["fqs"]);
@@ -566,7 +586,7 @@ export function splitMedications(
 export function useMedicationHistory(medication?: fhir4.MedicationStatement) {
   return useFeatureFlaggedQueryWithPatient(
     QUERY_KEY_MEDICATION_HISTORY,
-    [],
+    [medication?.id],
     "medications",
     "req.timing.medication_history",
     getMedicationHistoryFQS(medication),
@@ -594,7 +614,7 @@ export function useLastPrescriber(medication?: fhir4.MedicationStatement) {
         // 2. Throw away any resources that are not MedicationRequests.
         filter(propEq("resourceType", "MedicationRequest")),
         // 3. Sort by the authored on date.
-        sortBy((m) => Date.parse(m.authoredOn)),
+        fpSortBy((m) => Date.parse(m.authoredOn)),
         // 4. If there are med dispense records, make them models.
         map((mr) => new MedicationModel(mr, includedResources)),
         // 5. Take the last (latest) from our filtered list.
@@ -690,12 +710,15 @@ function getMedicationHistoryODS(medication?: fhir4.MedicationStatement) {
         ])
       );
 
-      const medicationResources = compact([
+      let medicationResources = compact([
         ...medicationStatementResponse.resources,
         ...medicationAdministrationResponse.resources,
         ...medicationRequestResponse.resources,
         ...medicationDispenseResponse.resources,
       ]).map((m) => new MedicationModel(m, includedResources));
+
+      // force ODS results to be sorted by id just like FQS results to ensure both functions return the same output.
+      medicationResources = sortBy(medicationResources, (a) => a.resource.id);
 
       const medications = sort(
         uniqWith(
@@ -753,12 +776,16 @@ function getMedicationHistoryFQS(medication?: fhir4.MedicationStatement) {
           resources.MedicationDispense
         ),
       ]);
-      const medicationResources = compact([
+      let medicationResources = compact([
         ...medicationStatementResponse.medications,
         ...medicationAdministrationResponse,
         ...medicationRequestResponse,
         ...medicationDispenseResponse,
       ]).map((m) => new MedicationModel(m));
+
+      // force FQS results to be sorted by id just like ODS results to ensure both functions return the same output.
+      // TODO: Remove once the ODS code path no longer exists.
+      medicationResources = sortBy(medicationResources, (a) => a.resource.id);
 
       const medications = sort(
         uniqWith(