+++ src/fhir/provenance.ts
@@ -98,6 +98,11 @@ export async function searchProvenances<T extends fhir4.Resource>(
   const targets = uniq(models.map((m) => `${m.resourceType}/${m.id}`));
   const ids = models.map((m) => m.id);
 
+  // Builder-scoped queries will fix the issue of FQS not supporting Practitioner provenance search.
+  if (enableFQS && models[0].resourceType === "Practitioner") {
+    return [];
+  }
+
   if (enableFQS) {
     return searchProvenancesFQS(requestContext, models[0].resourceType, targets);
   }
@@ -111,14 +116,20 @@ export async function searchProvenancesODS(
   patientUPID: string | undefined
 ) {
   const id = ids.join(",");
-  const { bundle } = await queryClient.fetchQuery([QUERY_KEY_PROVENANCE, id], async () =>
-    searchAllRecords(resourceType, requestContext, {
+  const { bundle } = await queryClient.fetchQuery([QUERY_KEY_PROVENANCE, id], async () => {
+    if (resourceType === "Practitioner") {
+      return searchAllRecords(resourceType, requestContext, {
+        _id: id,
+        _revinclude: "Provenance:target",
+      });
+    }
+    return searchAllRecords(resourceType, requestContext, {
       _id: id,
       _revinclude: "Provenance:target",
       // Need to include patient identifier to work around CPR.
       "patient.identifier": patientUPID ? `${SYSTEM_ZUS_UNIVERSAL_ID}|${patientUPID}` : "",
-    })
-  );
+    });
+  });
 
   return getResources(bundle, "Provenance");
 }