+++ src/services/fqs/queries/immunizations.ts
@@ -51,7 +51,16 @@ export const immunizationsQuery = gql`
           patient {
             reference
             resource {
-              ...Patient
+              ... on Patient {
+                ...Patient
+                managingOrganization {
+                  resource {
+                    ... on Organization {
+                      name
+                    }
+                  }
+                }
+              }
             }
           }
           occurrenceDateTime